+++ src/cmd.rs
@@ -25,15 +25,15 @@ use toml::Value;
 
 use exonum::blockchain::Service;
 use exonum::encoding::serialize::FromHex;
-use exonum::helpers::fabric::{
-    keys, Argument, CommandExtension, CommandName, Context, ServiceFactory,
+use exonum::helpers::{
+    fabric::{keys, Argument, CommandExtension, CommandName, Context, ServiceFactory}, Height,
 };
 use exonum::node::NodeConfig;
 
 use super::{gen_btc_keypair, AnchoringConfig, AnchoringNodeConfig, AnchoringRpcConfig};
 use details::btc::{self, transactions::FundingTx, PrivateKey, PublicKey};
 use details::rpc::{BitcoinRelay, RpcClient};
-use observer::AnchoringObserverConfig;
+use handler::observer::AnchoringObserverConfig;
 use service::AnchoringService;
 
 #[derive(Clone, Debug, Serialize, Deserialize)]
@@ -78,7 +78,7 @@ impl CommandExtension for GenerateNodeConfig {
                 "ANCHORING_OBSERVER_CHECK_INTERVAL",
                 false,
                 "This option enables anchoring chain observer with the given check interval \
-                 (in milliseconds).",
+                 (in blocks).",
                 None,
                 "anchoring-observer-check-interval",
                 false,
@@ -128,6 +128,11 @@ impl CommandExtension for GenerateNodeConfig {
         let observer_config = {
             let mut observer_config = AnchoringObserverConfig::default();
             if let Some(interval) = observer_check_interval {
+                ensure!(
+                    interval > Height::zero(),
+                    "`anchoring-observer-check-interval` should be greater than zero."
+                );
+
                 observer_config.enabled = true;
                 observer_config.check_interval = interval;
             }