+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Breaking changes
+
+- The anchoring chain observer logic has been moved to the `before_commit` stage. (#131)
+  Thus now `anchoring-observer-check-interval` is measured in blocks instead of milliseconds.
+
 ### Internal improvements
 
 - Added check that funding transaction in `anchoring-funding-txid` contains